chore: push recompilation for stdlib changes down to `fm`
