+++ crates/nargo/Cargo.toml
@@ -8,7 +8,6 @@ edition.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-dirs.workspace = true
 rustc_version = "0.4.0"
 build-data = "0.1.3"
